Renamed and deleted some files.
