+++ ui/src/app/computations/computation-card.tpl.html
@@ -16,4 +16,4 @@
 
 -->
 <div class="tb-uppercase" ng-if="item && parentCtl.types.id.nullUid === item.tenantId.id" translate>plugin.system</div>
-<div class="tb-uppercase" translate>{{item && item.state === 'CREATED' ? 'computation.created' : 'computation.suspended'}}</div>
+<div class="tb-uppercase" translate>{{'computation.created'}}</div>
