--- ui/src/app/computations/computations-fieldsetTest.tpl.html
@@ -1,36 +1,0 @@
-<!--
-
-    Copyright © 2016-2017 The Thingsboard Authors
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<md-button ng-click="onDeleteComputation({event: $event})"
-           ng-show="!isEdit && !isReadOnly"
-           class="md-raised md-primary">{{ 'computation.delete' | translate }}</md-button>
-
-<md-content class="md-padding" layout="column" style="overflow-x: hidden">
-    <fieldset ng-disabled="loading || !isEdit ">
-        <md-input-container class="md-block">
-            <label translate>computation.name</label>
-            <input required name="name" ng-model="computation.name">
-            <div ng-messages="theForm.name.$error">
-                <div translate ng-message="required">computation.name-required</div>
-            </div>
-        </md-input-container>
-        <md-input-container class="md-block">
-            <label translate>computation.description</label>
-            <textarea ng-model="computation.additionalInfo.description" rows="2"></textarea>
-        </md-input-container>
-    </fieldset>
-</md-content>
