--- ui/src/app/computations/computationsTest.tpl.html
@@ -1,23 +1,0 @@
-<!--
-
-    Copyright © 2016-2017 The Thingsboard Authors
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<tb-grid grid-configuration="vm.computationGridConfig">
-    <tb-computation-test computation="vm.grid.operatingItem()"
-        is-edit="vm.grid.detailsConfig.isDetailsEditMode"
-        the-form="vm.grid.detailsForm"
-        on-delete-computation="vm.grid.deleteItem(event, vm.grid.detailsConfig.currentItem)"></tb-computation-test>
-</tb-grid>
