+++ ui/src/app/computations/index.js
@@ -19,7 +19,7 @@ import uiRouter from 'angular-ui-router';
 import thingsboardApiComputation from '../api/computation.service';
 import thingsboardApiComputationJob from '../api/computation-job.service';
 import ComputationController from './computation.controller';
-import ComputationTestController from './computationsTest.controller';
+//import ComputationTestController from './computationsTest.controller';
 import ComputationJobController from './computation-job.controller';
 //import ComputationUploadDirective from './computation-upload.directive';
 import ComputationDirective from './computation.directive';
@@ -34,7 +34,7 @@ export default angular.module('thingsboard.computation', [
 ])
     .config(ComputationRoutes)
     .controller('ComputationController', ComputationController)
-    .controller('ComputationTestController', ComputationTestController)
+    //.controller('ComputationTestController', ComputationTestController)
     .controller('ComputationJobController', ComputationJobController)
     //.directive('fileModel', ComputationUploadDirective)
     .directive('tbComputation',ComputationDirective)
